Use NSInteger for button index values.

+++ Futility/UIActionSheet+FKBlocks.m
@@ -102,7 +102,7 @@ static void *kFKCancelBlockAssociationKey = (void *)&kFKCancelBlockAssociationKe
         if (dismissBlock != nil)
         {
             dispatch_async(dispatch_get_main_queue(), ^{
-                dismissBlock(actionSheet, (int)buttonIndex);
+                dismissBlock(actionSheet, buttonIndex);
             });
         }
     }
